// src/app/pages/auth/models/user.model.ts
// Generated by https://quicktype.io

export interface UserResponse extends UserCreate {
  id: number;
  role: string;
}

export interface UserCreate {
  name: string;
  email: string;
  password: string;
  avatar: string;
}

export interface UserUpdate {
  email: string;
  name: string;
}

export interface CheckEmail {
  email: string;
}

export interface CheckEmailResponse {
  isAvailable: boolean;
}